Make django-lifecycle much, much faster
Some of the work that the lifecycle mixin is during the initialization
of new model objects is very expensive and unnecessary. It's calculating
(and caching) field names and foreign key models per-object, rather than
per-class / model. All instances of a model are going to have the same
field names and foreign key model types so this work actually only needs
to be done once per model type.

Replacing cached methods with cached classmethods yields a very sizable
performance improvement when creating a bunch of new model instances.